NextJS has been the WORST tech stack I had ever come across my dev career.
NextJS killed this version of the project - the compilation times, speed, and waiting hours to only be able to realistically put out 10-15 minutes of work because it is such a resource hog is insane.

Just use Vite. We are in 2025.

In a desperate attempt, I did delete studio-penguino to hopefully help speed it up, tried all the optimization strategies in the book, to no avail.

8GB Ram should be sufficient enough to run any modern framework, I've had much more complicated setups that have had seconds of compilation times rather than minutes, AND they actually ran.

It's disrepectful, gut wrenching, and bring utter disgust out of me towards the NextJS community when the best response they can produce towards this is "Hurr Durr, 8GB ram ecks dee in 2025, GIT GUT".

AstroJS, AdonisJS, and Tanstack is the way,

If your framework needs to re-compile everything on each save, and excuses it on the hardware, using webpack in 2025, and some cheap solution instead of an existing FANTASTIC bunder that is Vite, it's a them problem.

Sure, server actions are amazing. But by the time even a 16GB ram rig compiles your changes, AdonisJS will be making LAPS around Next.

Which brings me to RSC, it looks great to everyone who's never had the pleasure to work with Inertia.

It is packed with potential security vulnerabilities waiting to happen. Not just from the users' side, but also from the framework side itself, and it has happened before.

Therre is a light at the end of the tunnel: uploadthing, shadcn, and betterAuth have been an absolute pleasure to work with. Fantastic libraries. Same with Prisma, though I've used it plenty.

NextJS was an absolute waste of time, and I am glad I got to use it for a personal project to really get to see why people dislike it that much.